renamed main.py to chess_text.py, shifted source files to src folder for organisation, added central main.py
